+++ lib/phauxth/authenticate.ex
@@ -2,20 +2,31 @@ defmodule Phauxth.Authenticate do
   @moduledoc """
   Authenticates the current user using sessions.
 
-  You need to set a `user_context` module in the config and define a
-  `get_by(%{"session_id" => session_id})` function (see the examples
-  section below) in this module.
-
-  See the documentation for Phauxth.Config for more information
-  about the `user_context`.
+  This module checks the current Plug session for a `phauxth_session_id`,
+  checks to see if the session id is valid and then sets the `current_user`
+  value accordingly.
 
   For information about customizing this Plug, see the documentation
   for Phauxth.Authenticate.Base.
 
-  ## Phauxth session authentication
+  ## Configuration / setup
 
-  This module checks the current Plug session for a `session_id`. It then
-  checks to see if the session id is valid.
+  Add the `user_context` module (the module you are using to handle
+  user data) to the config:
+
+      config :phauxth, user_context: MyApp.Accounts
+
+  The user_context module (in this case, MyApp.Accounts) needs to have a
+  `get_by(%{"session_id" => session_id})` function, which returns either
+  a user struct or nil.
+
+  In the example below, the Sessions.get_session/1 function retrieves the
+  data and checks if the session is still valid (has not expired, etc.).
+
+      def get_by(%{"session_id" => session_id}) do
+        with %Session{user_id: user_id} <- Sessions.get_session(session_id),
+        do: get_user(user_id)
+      end
 
   ## Options
 
@@ -31,20 +42,6 @@ defmodule Phauxth.Authenticate do
 
       plug Phauxth.Authenticate
 
-  Then, add the `user_context` module (the module you are using to handle
-  user data) to the config:
-
-      config :phauxth, user_context: MyApp.Accounts
-
-  Finally, define a `get_by(%{"session_id" => session_id})` function
-  in the `user_context` module (in this case, in MyApp.Accounts).
-
-      def get_by(%{"session_id" => session_id}) do
-        with %Session{user_id: user_id} <- Sessions.get_session(session_id),
-        do: get_user(user_id)
-      end
-
-  This function should return a user struct or nil.
   """
 
   use Phauxth.Authenticate.Base
@@ -56,8 +53,10 @@ defmodule Phauxth.Authenticate do
   """
   @spec add_session(Plug.Conn.t(), binary) :: Plug.Conn.t()
   def add_session(conn, session_id) do
+    IO.warn("Authenticate.add_session/2 will be removed in version 2.0.0")
+
     conn
-    |> put_session(:session_id, session_id)
+    |> put_session(:phauxth_session_id, session_id)
     |> configure_session(renew: true)
   end
 end