+++ test/support/session_helper.exs
@@ -17,7 +17,7 @@ defmodule Phauxth.SessionHelper do
     put_in(conn.secret_key_base, secret)
   end
 
-  def add_session(id, id_type \\ :session_id) do
+  def add_session(id, id_type \\ :phauxth_session_id) do
     conn(:get, "/") |> sign_conn() |> put_session(id_type, id)
   end